@@ -4,60 +4,60 @@
   "private": true,
   "homepage": "/account-management",
   "dependencies": {
-    "@commitlint/cli": "^17.1.2",
-    "@commitlint/config-conventional": "^17.1.0",
-    "@fortawesome/fontawesome-svg-core": "^1.3.0",
+    "@commitlint/cli": "^18.6.0",
+    "@commitlint/config-conventional": "^18.6.0",
+    "@fortawesome/fontawesome-svg-core": "^6.5.1",
     "@fortawesome/free-solid-svg-icons": "^6.0.0",
-    "@fortawesome/react-fontawesome": "^0.1.17",
+    "@fortawesome/react-fontawesome": "^0.2.0",
     "@tourmalinecore/react-table-responsive": "^0.5.0",
-    "@tourmalinecore/react-tc-auth": "^0.2.1-alpha.0",
+    "@tourmalinecore/react-tc-auth": "^1.0.0",
     "@tourmalinecore/react-tc-ui-kit": "^0.4.4",
     "axios": "^1.1.3",
-    "clsx": "^1.1.1",
+    "clsx": "^2.1.0",
     "cross-env": "^7.0.3",
     "history": "^5.3.0",
-    "husky": "^8.0.1",
+    "husky": "^9.0.6",
     "mobx": "^6.9.0",
-    "mobx-react-lite": "^3.4.3",
+    "mobx-react-lite": "^4.0.5",
     "moment": "^2.29.4",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-app-polyfill": "^3.0.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^6.5.0",
     "react-scripts": "^5.0.0",
-    "react-toastify": "^9.1.1",
+    "react-toastify": "^10.0.4",
     "sass": "^1.53.0",
-    "typescript": "^4.6.2",
-    "use-react-router-breadcrumbs": "^3.1.0",
-    "web-vitals": "^2.1.4"
+    "typescript": "^5.3.3",
+    "use-react-router-breadcrumbs": "^4.0.1",
+    "web-vitals": "^3.5.2"
   },
   "devDependencies": {
     "@types/history": "^5.0.0",
-    "@types/node": "^16.11.26",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.13",
-    "@typescript-eslint/eslint-plugin": "^5.13.0",
-    "@typescript-eslint/parser": "^5.13.0",
-    "api": "^4.5.2",
-    "cypress": "^12.10.0",
+    "@types/node": "^20.11.10",
+    "@types/react": "^18.2.48",
+    "@types/react-dom": "^18.2.18",
+    "@typescript-eslint/eslint-plugin": "^6.19.1",
+    "@typescript-eslint/parser": "^6.19.1",
+    "api": "^6.1.1",
+    "cypress": "^13.6.3",
     "cypress-mochawesome-reporter": "^3.3.0",
     "cypress-multi-reporters": "^1.6.1",
     "cypress-qase-reporter": "^1.4.2-alpha.2",
     "eslint": "^8.10.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-airbnb-typescript": "17.0.0",
+    "eslint-config-airbnb-typescript": "17.1.0",
     "eslint-plugin-cypress": "^2.12.1",
     "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-jsx-a11y": "^6.5.1",
     "eslint-plugin-react": "^7.29.3",
     "eslint-plugin-react-hooks": "^4.3.0",
     "npm-run-all": "^4.1.5",
-    "postcss-scss": "4.0.4",
+    "postcss-scss": "4.0.9",
     "standard-version": "^9.3.2",
-    "stylelint": "^14.5.3",
-    "stylelint-config-standard": "^25.0.0",
-    "stylelint-order": "^5.0.0",
-    "stylelint-scss": "^4.1.0"
+    "stylelint": "^16.2.0",
+    "stylelint-config-standard": "^36.0.0",
+    "stylelint-order": "^6.0.4",
+    "stylelint-scss": "^6.1.0"
   },
   "scripts": {
     "start": "run-s create-config:dev start-devserver",
@@ -80,6 +80,10 @@
     "cy:open": "cypress open --component --browser chrome",
     "cy:e2e": "npx cypress run -e2e",
     "cy:component": "npx cypress run --component"
+  },
+  "peerDependencies": {
+    "react": "^17.0.0 || ^18.0.0",
+    "react-dom": "^17.0.0 || ^18.0.0"
   },
   "browserslist": {
     "production": [
